Read a file with crosswalk information between the blocks of two consecutive decennial censuses.

Read a file with crosswalk information between the blocks of two consecutive decennial censuses.

Usage

read_crosswalk(.filepath, .newer_vintage, ...)

Arguments

.filepath

<chr> the full path to the crosswalk file, which may be a ZIP archive

.newer_vintage

<int> The more recent of the two decennial censuses being compared.

...

<dynamic-dots> Additional arguments for readr::read_delim().

Value

<tbl> a many-to-many relationship table between the two census's blocks.

See also