Create a tigershape filename based upon state, year, and geography

Create a tigershape filename based upon state, year, and geography

Usage

tiger_name_for(
  state,
  year,
  geography = c("county subunits", "school districts", "places", "tracts", "blocks")
)

Arguments

state

the two-digit FIPS number for the state

year

the four-digit census year

geography

<chr, optional> one of "county subunits" (the default), "school districts", "places", "tracts", "blocks"

Value

a string such as "tl_2000_55_cousub00"

Examples

tiger_name_for(55L, 2000L, "county subunits")
#> [1] "tl_2000_55_cousub00"